pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "infrastructure-security-auditor"
version = "0.1.0"
description = "Automated Windows infrastructure security auditing tool with AI-powered recommendations"
authors = [
    { name = "Infrastructure Security Auditor Team" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["security", "audit", "windows", "infrastructure", "compliance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "anthropic>=0.25.0",
    "pywinrm>=0.4.3",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "flake8>=7.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[project.scripts]
auditor = "src.cli:cli"

[project.urls]
Repository = "https://github.com/your-org/infrastructure-security-auditor"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "setup.py"]

[tool.coverage.report]
show_missing = true
fail_under = 80
